requirements.txt
anyio==4.8.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.1.31
charset-normalizer==3.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
python-dotenv==1.0.1
python-telegram-bot==13.15
pytz==2025.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tornado==6.1
typing_extensions==4.12.2
tzlocal==5.3
urllib3==2.3.0
